Words connected to database; altered styling and statefulness
Styling changed - reduced / altered use of CSS file, more reliance on
inline styles dependent on state.
Buttons (actually divs) added for user to choose option (of two words).
(Placeholder) feedback given on click.
Commented out old code (should have updated this is a separate commit)
flipped through pairs in a list to give button word values. Current
code run as-is gets the words from the database.
Pressing “progress” now flips between word pairs. This must be done
before the buttons are pressed as I haven’t yet put in a mechanism to
return to initial state.
More use of state in the code. (TODO: move the state up to the highest
possible level in the object hierarchy - simpler and corresponds to
best practice.)
Attempts to get sounds from the database and play them at the right
time have so far failed. I seem to be able to get the objects from the
database, but not to play them.