button linkage
Progress button and option buttons now disable / enable one another.
Part of the way this was done was kind of dodgy - React actually says
that there is an error, but it sort of counts more as a warning. The
problem is that I’m setting state inside the render function of the
option buttons, which is not very good practice as it could lead to
infinite loops. Also, the way I’ve done the counter sometimes means it
takes two clicks to activate a button.
Added more words to database, so now there are 4 option buttons. Yet to
sort this out.
The reason I’m uploading such dirty code is because I’m about to
completely change the way state works, for the nth time. Now I think I
may have finally cracked how to get the Arena to be in charge via state
(only got halfway there so far)…