Primitive connection of buttons to sounds (still buggy)
A random sound plays on button press.
One option is highlighted as “correct”, the others as “wrong”, as soon
as an option is clicked. This is based on what sound was played.
Unfortunately, it seems to be inconsistent as it sometimes highlights
the wrong option (out of “thing” and “sing”) as correct.