Files now autofill "speaker" and "Item" fields on save
When you save, the program splits the file name into two parts: the
‘speaker’ part (e.g. “wozza”) taken from the beginning, and the ‘item’
part (e.g. “week”) taken from the end. Then it iterates through the
items’ homophones, and if it finds a match, then it submits the object
id of the item where it found a match and makes a pointer to that item.