Item submission duplicate checking
When you submit an item (via homophones),  the program loops through
the items checking for duplicates, and only allows the new item to be
added if no duplicated are found. If duplicates are found, an error is
logged, and we are told what the objectId and homophones of the
existing item are.