Arena page Progress button functions
Now the “Progress” button’s label and actions depend on what state the
Arena is in.
In “ask”, “wait”, and “done” mode, it chooses a new pair and play a
sound. In “done” mode, it additionally resets the score and the counter.
In “feedback” mode, it replays the current sound.
(Also, added a new state called “done”, when the user has gone through
the whole of training.)