Dummy dispatchAudio function
Made firstNodes into a list of words *and ids* so that the ids can be
submitted with the words when we do a mutation.
The dispatchAudio function asks for the user’s name by prompting with a
dialog box, and then calls submitAudio for every element in a zipped
array of items, their ids, and the corresponding audioURLs. (Maybe will
have to do some more work to make this blobs rather than URLs.)