Footer background-color removed
I removed the footer's background-color.

Also this is a test for the Github application on my computer. It has a
Sync button, which says: "Pull changes from and push local changes on
master to the remote"
But the last time I used this button, you didn't see it. So now I commit
this, than I will use the Sync button to see, if you can see the
changes.