Graph width, access denial for guests to some pages, changes to Arena styling
Graphs on profile page now full width and resize responsively with the
window.
Guests are now not permitted to use the record page, but are asked to
log in or sign up.
For development purposes, whether the user is seen as logged in or not
is based on a hard-coded variable. This allows us to program both the
experience of a logged-in user and that of a guest by changing the value
of this variable (in app.jsx).
The username in the nav is no longer a button, as it was considered
unnecessary. Logged-in users will now view their profile in the home
page.
Arena colours changed: now progress fluid is blue, progress bar is
orange. Sizing changed to stop scrollbar appearing. Score display
altered and centered.
Home page buttons text styling changed to vw (viewport-width) units.
This allows font to scale with the viewport so that the text remains
within the buttons on any screen size.
Dropdown menu from gears "fixed" - it was not appearing below the nav,
but on top of it.