Login page change; Contact Us; more links
The Moderation page is not necessary and has been removed (as distinct
from the Edit page, which is what we are going with).
Now there is a link to the Edit and Contact pages.
When you log in, you are sent over to the home page.
The Nav bar dropdown now has Moderation and Contact Us as additional
options.
The reason for the uneven text inputs on the log in page was an extra
space (i.e. spacebar!) in the code - now that that has been removed, it
is nice and symmetrical.