Ask for native language; Navbar changes; keyboard event; suggestions page
- When user registers, they must provide their native language, so that
we know what record page to show them (the record page for which
language).
- Log-in now works by pressing the enter button when the password input
box is in focus.
- New suggestions page: for users to suggest new contrasts to add to
existing languages, or new word pairs to add to existing contrasts. Not
yet fleshed out.
- Added Register (sign up) and Suggest words to the Nav dropdown menu
- Nav dropdown moves together with the navbar: previously if you
scrolled down, the dropdown would stay in absolute position (so that
you would scroll down the dropdown too, and it would disappear away
from you at the top of the page).