Set up logo as home button
Now the image is used as the home button.
At the moment this is the same as the favicon. This can be easily
swapped out - the code would be the same, just the src of the img
element would change.

Also, all the panels now fade in (previously some were and some
weren’t).