View Stats as second button available when training ends
One of the pieces of feedback we got from users of Alpha is that there
should be more than one option (“Play Again”) when you end - like Main
Menu. Now there is one - View Stats (essentially like Main Menu, since
we technically don’t have a main menu, and the stats are shown on the
profile page, which is the home page if you are logged in).