About page PDFs
PDFs have been put on the server, and the About page now links to them.
Also some text changed and some links added on the About page.
Comment on the bottom gives the source of the PDFs for reference
purposes. The Amazon links go stale with time, but I thought I may as
well leave them anyway.