Tutorial button fixed
Tutorial button is now working - it plays the tutorial again after you
have gone through it.

Other changes:
- added more steps to the tutorial to explain the other features of the
record page
- added new, edited versions of the quack wrong and the trumpet at the
end of training to be quieter and therefore less unpleasant/annoying,
and more in line with other volumes
- re-introduced the trumpet at the end of training
- arena waits a moment before transitioning to results page, so that
you can see the progress bar reach the end (previously it was going
straight to the results page and so you didn’t ever see the progress
bar reach the end)
- changed var to let or const where appropriate in arena

Remaining bug:
- The tutorial doesn’t switch the text from language to language when
you set locale on the record page, you have to move to another page and
come back. I have documented this bug in a long comment and understand
its origin but am undecided on the best way to fix it.