New app-level state: interfaceLanguage
This solves the problem from the previous commit concerning not having
some things update the language because they were using
counterpart.translate (JavaScript function) rather than <Translate>
(React object designed to pick up changes in counterpart).

It seems a little heavy-handed to made a top-level state to deal with
this, but on the other hand we have a top-level state to deal with
which language the user has decided to be training, so perhaps it isn’t
so bad.