Average Contrast statistic taken from database
DonePanel also separated from Arena into a different file.
At the moment the averageContrast doesn’t display because it is null,
but the console.log shows that it is being fetched without errors (it’s
just that at this moment in time we don’t have any value in the
database for this).