allStats query and completeTutorial mutation
allStats query works, but there is currently no data to display.
tutorialMutation query *almost* works (client-side it is calling at the
right time, and its syntax is being understood), it just appears to be
throwing a type error at the moment, which might have something to do
with the backend.