BetaSign; display message when no training in period
BetaSign: the icon has BETA written on it (for Beta release).
In Charts, when there is no data for a given period (e.g. if the user
has trained in the past year but not in the past week, and wants to
view graphs for the past year), display a message to this effect rather
than empty charts.