Tiny graphql error fix
We have a choice atm between a graphql error that doesn’t seem to do
any harm, or a tiny glitch for the users. The previous commit has the
former; this has the latter. My hope is that there is a way to get rid
of both, but in the interests of making sure the app is running well,
I’m taking the removal of the error as a higher priority than the very
slightly less appealing behaviour for users.
When the Record page is loaded, there are two wrappers waiting for
queries to return. While waiting, each returns the LoadingPage.
Previously, both the calls of LoadingPage were inside the inner
wrapper, so that there was no transition for the users, but there was a
gql error as it was trying to execute the second query before the first
had returned (and the first provides props for the second). Now, this
is solved as the LoadingPage is called twice (once by each wrapper),
but it means that the users see the LoadingPage start twice, which
doesn’t look as good.
Hopefully future commits will resolve this. It’s not bad enough to keep
out of Master. I’m leaving it as a todo.