Unsuccessful attempt due to graphql returning a function not a React element
React wants render() to return a React element, but gql returns a
function that returns a React element based on props and context. So
this is stopping this from working at the moment, as the idea was to
decide whether or not to wrap the app with the query for account info
depending on the top-level state of being logged in or not. The
wrapping is not working.