Return old Record structure; Move order of actions in the app setUser method; remove event listener if Nav recreates
Old Record page structure is required - the new one was based on
dealing with user info on the Record page level (rather than the app
level).
localeStorage.setItem(“token”, raw_jwt) before this.setState({
isLoggedIn: true )} in app.jsx.
Nav destroys event listener in window for clicks to close the dropdown
after the Nav itself unmounts (as a different Nav will mount, depending
on whether the user is logged in or not).