Record page refreshing on submission
Reset the state and refetch the query of items. (Though the latter
appears to be giving the same items every time at the moment.)
Also, whether tutorial is shown or not, and whether pre-record page
displays or not, is based on whether the user has seen it before (which
is taken from the database).