NoEditPage for non-moderators (client-side only)
Page that shows if you’re not a moderator and navigate to the
moderation page. (That is, if you type in the url for the moderation
page. The button to take you there would be absent in the nav dropdown
if you’re not a moderator.)
Needs to be based on queries, as of yet it isn’t.