Fix localisation code; fix sql bug in function names
There was a function being referred to that didn't exist in
schema-private.sql - now it refers to the previously renamed function.
The localisation code was returning undefined as a locale on start-up
due to a faulty ternary if statement. This was fixed, and the code was
improved:
- now accepts 2-letter codes with and without country codes, as well
as 3-letter codes
- the big object localeMapper used to look these up is now much smaller