Fix nav bug; repeatedly clicking gears toggles menu rather than keeping it in place
Bug was due to not using state in one place:
this.mouseIsDownOnDropdown -> this.state.mouseIsDownOnDropdown.
Toggling between down and not down when clicking on the gears seems
much more natural. Did this by using onMouseDown and onMouseUp with the
gears image <li> (as with the dropdown itself), and then allowing a
follow-up function to be passed, which toggles the dropdown on or off.