Add languages from ISO-639-3 to data.sql
https://en.wikipedia.org/wiki/List_of_ISO_639-2_codes

The intention was to use ISO-639-2, but in some cases these have more
than one code. ISO-639-3 has the same codes ISO-639-2, but
disambiguates to a single code where ISO-639-2 has two codes.

Languages with a writing system that couldn’t be read, or with no
native name listed, were removed from the list.
Chinese as a macro-language was removed; Mandarin and Cantonese were
inserted, but other Chinese languages are not in the list.