bugfix: `this` in Record page; zip util function
We were previously using _ (lodash / underscore) for zipping. But we
were only using that library in one line in the entire codebase. Better
that we have our own zip util function.

…unless some of our other dependencies require lodash / underscore, so
it’s available anyway? In which case we could use zip. I’m replacing it
with our own util here now though.